В TimeFormatter изменил тип входящего аргумента, поправил шрифты, настроил степперы для таймеров раунда и отдыха

+++ Timer/TimeFormatter.swift
@@ -11,13 +11,13 @@ import Foundation
 ///Класс для вывода времени в формате MM:ss:msms
 class TimeFormatter {
     
-    static func formatter(time: Float) -> String {
+    static func formatter(time: Double) -> String {
         
         let mm = String(format: "%02d", Int(time / 60))
         let ss = String(format: "%02d", Int(time.truncatingRemainder(dividingBy: 60)))
 
         let b = Int(time)
-        let afterPoint = time - Float(b)
+        let afterPoint = time - Double(b)
         let mls = String(format: "%02d", Int(afterPoint * 100))
         
         return "\(mm):\(ss):\(mls)"
